crewAI/chapter2/requirements.txt
crewai
langchain-ollama
python-dotenv
duckduckgo-search
